chore(refactor): moved fixEnv til servires from core
