+++ packages/cli/src/services/processComparisonFile.ts
@@ -3,7 +3,7 @@ import { parseEnvFile } from './parseEnvFile.js';
 import { filterIgnoredKeys } from '../core/helpers/filterIgnoredKeys.js';
 import { compareWithEnvFiles } from '../core/scan/compareScan.js';
 import { findDuplicateKeys } from '../core/duplicates.js';
-import { applyFixes } from '../core/fixEnv.js';
+import { applyFixes } from './fixEnv.js';
 import { toUpperSnakeCase } from '../core/helpers/toUpperSnakeCase.js';
 import { resolveFromCwd } from '../core/helpers/resolveFromCwd.js';
 import { detectEnvExpirations } from './detectEnvExpirations.js';
